refactor(purchaseSubmit): Move purchaseSubmit into it's own utility file
