+++ client/components/Stocks.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import {connect} from 'react-redux'
-import axios from 'axios'
 import {me, updateShares} from '../store'
-import {toastr} from 'react-redux-toastr'
+import {purchaseSubmit} from '../utilities'
 
 const Stocks = props => {
   const {
@@ -41,35 +40,6 @@ const Stocks = props => {
       ))}
     </section>
   )
-}
-
-const purchaseSubmit = (
-  stockInfo,
-  userId,
-  balance,
-  fetchBalance,
-  updatePortfolio
-) => evt => {
-  evt.preventDefault()
-  const shares = Number(evt.target.shares.value)
-  const total = stockInfo.price * shares
-  if (total < balance) {
-    const objToSend = {
-      shares,
-      pricePerShare: stockInfo.price,
-      type: 'purchase',
-      symbol: stockInfo.symbol,
-      userId
-    }
-    axios
-      .post('/api/transactions', objToSend)
-      .then(res => res.data)
-      .then(updatePortfolio)
-      .then(() => fetchBalance())
-      .catch(console.log)
-  } else {
-    toastr.warning('Insufficient Funds', 'Please deposit more funds into your account') // change this to some sort of state
-  }
 }
 
 const mapStateToProps = state => ({
