+++ client/utilities/index.js
@@ -1,1 +1,2 @@
 export {default as updateSharesHelper} from './updateSharesHelper'
+export {default as purchaseSubmit} from './purchaseSubmit'
